Started fixing bugs on G functions of LSH

+++ parsing/parsingLSH.h
@@ -3,7 +3,7 @@
 
 
 void readFile(char*,List * ,int *);
-void readQueryFile(char*,char*,LSH,List,int,int );
+void readQueryFile(char*,char*,LSH,List,int,double );
 int findDim(char* );
 
 #endif
